+++ lambda-notes/src/components/Notes/NewNote.js
@@ -1,7 +1,7 @@
 import React from "react";
 import axios from "axios";
 import {Link} from "react-router-dom";
-import "../App.css";
+import "./Notes.css";
 
 class NewNote extends React.Component {
   constructor(props) {
