+++ lambda-notes/src/components/Notes/NoteCard.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../App.css";
+import "../../App.css";
 import {Link} from "react-router-dom";
 
 const NoteCard = props => {
